mindful_moments/requirements.txt
Flask==2.3.3
mysql-connector-python==8.3.0
requests==2.31.0
